restructuring single-cell data icon exports for landing page
