+++ frontend/src/pages/landing-page/icons/single-cell-data-inactive.tsx
@@ -1,0 +1,76 @@
+const SingleCellDataIconInactive = (): JSX.Element => (
+  <svg
+    width="48"
+    height="55"
+    viewBox="0 0 48 55"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M12.4725 4.37666L13.1236 5.24472L12.4725 4.37666ZM46.3826 26.0801L45.3023 26.183L46.3826 26.0801ZM21.5153 53.5843L21.4816 52.4997L21.5153 53.5843ZM1.54581 43.7126L2.62183 43.8529L1.54581 43.7126ZM13.1236 5.24472C16.6701 2.58454 20.317 1.91872 23.8392 2.53514C27.3988 3.15811 30.895 5.10515 34.0287 7.78155C37.1568 10.4531 39.8678 13.8047 41.8557 17.1383C43.8539 20.4893 45.0691 23.7337 45.3023 26.183L47.4628 25.9772C47.1874 23.0853 45.8052 19.5242 43.7197 16.0268C41.6238 12.512 38.7646 8.97227 35.4382 6.13126C32.1174 3.29505 28.2747 1.10815 24.2133 0.397362C20.1144 -0.319982 15.8581 0.480633 11.8213 3.50859L13.1236 5.24472ZM45.3023 26.183C46.6016 39.825 36.5203 52.0327 21.4816 52.4997L21.5489 54.6689C37.8434 54.1629 48.8806 40.8636 47.4628 25.9772L45.3023 26.183ZM21.4816 52.4997C15.5376 52.6843 10.5494 51.6731 7.18602 49.9827C3.78536 48.2735 2.33429 46.0574 2.62183 43.8529L0.469788 43.5723C0.00518179 47.1343 2.47249 50.0426 6.21141 51.9218C9.98763 53.8197 15.3602 54.8611 21.5489 54.6689L21.4816 52.4997ZM2.62183 43.8529C2.82005 42.3332 3.447 40.9462 4.33783 39.5479C5.24222 38.1283 6.35692 36.7849 7.52841 35.2701C9.82755 32.2972 12.2012 28.8438 12.2012 23.9694H10.0309C10.0309 28.0683 8.07157 31.0202 5.81165 33.9424C4.70361 35.3751 3.48225 36.8517 2.50745 38.3818C1.51908 39.9332 0.723692 41.6256 0.469788 43.5723L2.62183 43.8529ZM12.2012 23.9694C12.2012 22.1318 11.9051 20.2322 11.5752 18.4058C11.2393 16.5467 10.874 14.788 10.6842 13.0939C10.306 9.71802 10.6832 7.07526 13.1236 5.24472L11.8213 3.50859C8.42589 6.05548 8.12489 9.74262 8.52743 13.3355C8.72801 15.1258 9.11782 17.0112 9.43948 18.7917C9.76707 20.6049 10.0309 22.3405 10.0309 23.9694H12.2012Z"
+      fill="#0073ff"
+    />
+    <circle
+      cx="18.9905"
+      cy="17.2587"
+      r="2.03462"
+      stroke="#0073ff"
+      strokeWidth="2.17027"
+    />
+    <circle
+      cx="14.6951"
+      cy="36.4035"
+      r="2.03462"
+      stroke="#0073ff"
+      strokeWidth="2.17027"
+    />
+    <circle cx="17.6718" cy="48.4818" r="1.31874" fill="#0073ff" />
+    <circle cx="9.50752" cy="38.8905" r="1.31874" fill="#0073ff" />
+    <circle cx="16.7298" cy="26.6825" r="1.31874" fill="#0073ff" />
+    <circle cx="42.3133" cy="31.0187" r="1.31874" fill="#0073ff" />
+    <circle cx="33.8651" cy="38.4381" r="1.31874" fill="#0073ff" />
+    <circle cx="24.1148" cy="31.8041" r="1.31874" fill="#0073ff" />
+    <circle cx="35.9833" cy="17.2079" r="1.31874" fill="#0073ff" />
+    <circle cx="16.9559" cy="9.01161" r="1.31874" fill="#0073ff" />
+    <circle
+      cx="39.7889"
+      cy="23.3291"
+      r="2.03462"
+      stroke="#0073ff"
+      strokeWidth="2.17027"
+    />
+    <circle
+      cx="28.6233"
+      cy="35.5371"
+      r="2.03462"
+      stroke="#0073ff"
+      strokeWidth="2.17027"
+    />
+    <path
+      d="M25.8505 8.44126C25.949 7.8501 25.5496 7.29104 24.9585 7.19256C24.3673 7.09408 23.8082 7.49346 23.7098 8.08462L25.8505 8.44126ZM28.7448 13.4482C29.3404 13.3811 29.7687 12.8439 29.7016 12.2484C29.6345 11.6528 29.0973 11.2245 28.5018 11.2916L28.7448 13.4482ZM23.7098 8.08462C23.4736 9.50235 23.8322 10.9372 24.7352 11.9854C25.6566 13.055 27.0725 13.6366 28.7448 13.4482L28.5018 11.2916C27.4991 11.4045 26.8051 11.0631 26.3795 10.569C25.9355 10.0535 25.7101 9.28422 25.8505 8.44126L23.7098 8.08462Z"
+      fill="#0073ff"
+    />
+    <path
+      d="M26 19.979C25.4008 19.9646 24.9035 20.4386 24.8891 21.0378C24.8747 21.6369 25.3488 22.1343 25.9479 22.1486L26 19.979ZM30.5501 16.4094C30.4 15.8292 29.8079 15.4806 29.2277 15.6308C28.6475 15.781 28.2989 16.3731 28.4491 16.9532L30.5501 16.4094ZM25.9479 22.1486C27.3847 22.1831 28.7549 21.6263 29.6657 20.5848C30.5951 19.5222 30.9719 18.0386 30.5501 16.4094L28.4491 16.9532C28.702 17.9301 28.4615 18.6652 28.0321 19.1561C27.5843 19.6681 26.8543 19.9995 26 19.979L25.9479 22.1486Z"
+      fill="#0073ff"
+    />
+    <path
+      d="M33.3336 35.0276C33.2239 35.6168 33.6127 36.1834 34.2018 36.293C34.791 36.4027 35.3575 36.014 35.4672 35.4248L33.3336 35.0276ZM30.5348 29.9667C29.9381 30.0225 29.4996 30.5515 29.5554 31.1482C29.6112 31.7449 30.1402 32.1834 30.7369 32.1276L30.5348 29.9667ZM35.4672 35.4248C35.7303 34.0118 35.3989 32.5704 34.5159 31.5053C33.615 30.4184 32.2104 29.81 30.5348 29.9667L30.7369 32.1276C31.7415 32.0336 32.4289 32.3882 32.8451 32.8903C33.2792 33.414 33.49 34.1875 33.3336 35.0276L35.4672 35.4248Z"
+      fill="#0073ff"
+    />
+    <path
+      d="M11.5318 47.4897C11.6129 48.0835 12.16 48.4991 12.7538 48.418C13.3476 48.3368 13.7632 47.7897 13.6821 47.1959L11.5318 47.4897ZM7.28369 43.5654C6.7348 43.806 6.48486 44.446 6.72543 44.9949C6.96601 45.5438 7.606 45.7937 8.15489 45.5531L7.28369 43.5654ZM13.6821 47.1959C13.4875 45.7719 12.7197 44.5078 11.5467 43.7742C10.3497 43.0257 8.82508 42.8898 7.28369 43.5654L8.15489 45.5531C9.07903 45.1481 9.84307 45.2686 10.396 45.6143C10.9728 45.975 11.4161 46.643 11.5318 47.4897L13.6821 47.1959Z"
+      fill="#0073ff"
+    />
+    <path
+      d="M22.1656 43.0818C22.1742 42.4826 21.6954 41.9898 21.0961 41.9813C20.4969 41.9727 20.0041 42.4515 19.9956 43.0507L22.1656 43.0818ZM25.7791 47.5972C26.3578 47.4414 26.7006 46.846 26.5448 46.2673C26.389 45.6886 25.7936 45.3458 25.2149 45.5016L25.7791 47.5972ZM19.9956 43.0507C19.975 44.4879 20.5451 45.8526 21.5953 46.7533C22.6668 47.6723 24.154 48.0347 25.7791 47.5972L25.2149 45.5016C24.2406 45.7639 23.5032 45.5305 23.0082 45.1059C22.4918 44.663 22.1534 43.9363 22.1656 43.0818L19.9956 43.0507Z"
+      fill="#0073ff"
+    />
+    <path
+      d="M29.8042 25.3637L29.9352 24.2865L29.9032 24.2827L29.8712 24.2807L29.8042 25.3637ZM19.7816 31.9199L18.6982 31.8597L19.7816 31.9199ZM29.6732 26.4409C31.9566 26.7186 33.9609 28.6256 35.4324 31.0508C36.1506 32.2345 36.7007 33.4776 37.0599 34.5718C37.4288 35.6954 37.5635 36.5622 37.5374 37.0197L39.7041 37.1435C39.7533 36.2823 39.5207 35.1097 39.1219 33.8949C38.7135 32.6507 38.0955 31.2563 37.2878 29.925C35.7074 27.3202 33.228 24.687 29.9352 24.2865L29.6732 26.4409ZM37.5374 37.0197C37.4825 37.98 36.726 39.5384 35.4685 40.8662C34.2079 42.1974 32.7615 42.9651 31.5579 42.8582L31.3658 45.0199C33.5533 45.2143 35.611 43.8721 37.0443 42.3585C38.4809 40.8417 39.6083 38.8206 39.7041 37.1435L37.5374 37.0197ZM31.5579 42.8582C30.9964 42.8082 30.0216 42.3967 28.7731 41.5737C27.5699 40.7806 26.2498 39.698 25.0302 38.4984C23.8089 37.2971 22.7179 36.0072 21.9536 34.8093C21.161 33.567 20.8309 32.5949 20.8651 31.9801L18.6982 31.8597C18.6245 33.1857 19.2767 34.6486 20.124 35.9767C20.9997 37.3492 22.2064 38.765 23.5083 40.0456C24.8119 41.3279 26.2403 42.5035 27.5787 43.3857C28.8719 44.2381 30.2318 44.9191 31.3658 45.0199L31.5579 42.8582ZM20.8651 31.9801C20.9476 30.4945 22.144 29.0627 24.0468 27.9717C25.9195 26.8981 28.1596 26.3492 29.7371 26.4468L29.8712 24.2807C27.7938 24.1521 25.1357 24.8458 22.9674 26.089C20.8292 27.3148 18.8418 29.2742 18.6982 31.8597L20.8651 31.9801Z"
+      fill="#0073ff"
+    />
+  </svg>
+);
+
+export default SingleCellDataIconInactive;
